Revert "fix lsrro circular import (#1700)"
This reverts commit c8da2e840533ba5866473f40198448d41ca089e3.
